moviesDescription v0.1

+++ lib/models/salle.dart
@@ -1,14 +1,14 @@
 class Salle{
   final int id;
   final String name;
-  final double nbrPlaces;
+  final int nbrPlaces;
 
   Salle({this.id,this.name, this.nbrPlaces});
   
   factory Salle.fromJson(Map<String, dynamic> json) {
     return Salle(
       id: json['id'] as int,
-      nbrPlaces: json['nombrePlaces'] as double,
+      nbrPlaces: json['nombrePlaces'] as int,
       name: json['name'] as String,
     );
   }
